terminal_terrestre/requirements.txt
# ============================================================
# Terminal Terrestre — Dependencias del Backend
# ============================================================

# Framework principal
Django==5.0.6

# DRF — solo para endpoint API de validación QR externa
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1

# Base de datos
psycopg2-binary==2.9.9

# CORS
django-cors-headers==4.4.0

# Generación de códigos QR
qrcode[pil]==7.4.2
Pillow==10.4.0

# Peticiones HTTP (integración Hikvision ISAPI)
requests==2.32.3

# Variables de entorno
python-dotenv==1.0.1

# Servidor WSGI para producción en Windows
waitress==3.0.0

# Total en letras (reporte de recaudación)
num2words==0.5.13
